feat(search): setups variation B of filters
