+++ src/components/advanced-search/index.tsx
@@ -1,4 +1,4 @@
-import SearchFilters from "#/components/advanced-search/search-filters";
+import { SearchFilters } from "#/components/advanced-search/search-filters";
 import type { IParams } from "#/models/search/search-filter-params";
 import styles from "./style.module.css";
