+++ src/components/advanced-search/style.module.css
@@ -14,6 +14,7 @@
 .filter-container {
   display: flex;
   flex-wrap: wrap;
+  align-items: center;
 }
 
 .map-switch {
@@ -28,4 +29,37 @@
     width: 100%;
     max-width: 100%;
   }
+}
+
+.etatSelect {
+  width: 110px;
+  height: 32px;
+  padding: 0 18px;
+  font-size: 1rem;
+  line-height: 1.5rem;
+  color: var(--text-default-grey);
+  text-align: center;
+  text-align-last: center;
+  appearance: none;
+  background-color: transparent;
+  border: 1px solid #ddd;
+  border-radius: 50px;
+  box-shadow: none;
+}
+
+.etatSelect[data-etat="A"] {
+  width: 120px;
+}
+
+.etatSelect[data-etat="C"] {
+  width: 92px;
+}
+
+.etatSelect:hover {
+  border-color: #b5b5b5;
+}
+
+.etatSelect:focus {
+  outline: 2px solid var(--annuaire-colors-frBlue);
+  outline-offset: 2px;
 }
